Finished Angle Force Control Simple Spring -  Having Little Trouble with Angle Calculations
Everything is just about set for the simple spring simulation.  I'm
having some trouble with applying the force at the right angle
consistently as the foot moves in the various quadrents.   There's
something I need to fix with the trig, but I'm not figuring it out right
now.

Also added a force indicator sprite, but haven't done any of the scripts
for it.