FIXED ANGLE ON LEG!  + Made 1D Version of Real Leg Design
After sitting down and doing the math, like 5 times, I finally was able
to get the math properly, and the leg works!  Yay!  I still have a
little weird on the foot control, so I'll give it an option to invert
how the foot is moved.

I also put in a bunch of hours to set up a real leg design.  It seems,
by the simulation so far, that a 1D version of our leg should work,
given the right spring.  So we'll have to play with that IRL.   Overall,
good work today!