Set up torque on body
I figured out how to find if the foot was touching any of the walls  or
anything in the space (although not in the best way) in the foot control
script (on the hip gameobject), and was able to set up the foot control
script to apply a force to the body if the foot is on the ground at all.
This should allow for better mobility and control.