Created Clamp + Cleaned Code
I did some changes to create the clamp ability.  I think it'll do some
good things, personally, once it's set.  It'll be cool.  I also realized
I had a bunch of unnecessary code that I cleaned up and reorganized to
make it easier to interact with.  Still need to fix up some bugs and
clean a little bit, but good work so far.