Finished Clamp
Found four conditions of when to clamp. When it's retracted. When the
distance between the foot and body isn't changing. When the distance
between the foot and body is greater than the spring's set distance
(meaning there's something stuck within it). When the foot and the body
have 0 velocity.

There was still an issue with that last condition because it would
register the clamp as happening RIGHT when I retract if the foot and
body were at rest when I hit e.

I didn't solve this issue, but I got around the bug it caused by
increasing the frequency when clamped at a certain rate rather than
instantly. This causes the foot and body to come to a rest more easily
rather than shooting up because of the high spring frequency.