Fixed Clamp + Made more in-depth Moving Block
Worked on clamp conditionals a little bit that let me register clamp on
a moving object.

The majority, though, is the work I did on the Moving Platform script.
I made it so that you can stay in place and define a path for it to
follow (only horiz or vert for now).  It can also repediately follow the
path or loop back on itself.  I think I'm almost set to build some basic
levels.