Finished Text, Updated Levels, Finished Level 4
Basically, what the title says.  I finished and went back to the levels
to polish them up and put in the new fading in text.  Also, added them
to the build, so that they go from one to the next automatically now.

Finished level 4.  Should probably add some text to it to give some
hints, but we'll see what people think first.