Finsihed Hazard, Explosion and Reset Level
I figured out what was wrong with the explosion script.  I got a working
explosion, and reset and moved around and added a bunch of things in
order to create the explosion and reset the level once a player hits a
hazard.  It works well now.  I can create some hazard levels.  And
possibly some enemies. :)