Implemented Checkpoints, Updated Player Prefab
Checkpoints work well now.  A copy of the player spawns at the
checkpoint when you die or resets the level if no checkpoints were
active.

Also cleaned up a bit on the player prefab because it was bugging me and
it was giving errors in the scipting when creating a new player.