Fixed Box Maker, Added Rand Force, Updated Action Switch + Fixed Bug
Fixed the Box Maker to make it seperate and run on an action switch.

Also did a lot of work with the action switch to make it work and fix a
but I didn't know I was having regarding pressing the control.   I
thought it was just my keyboard.  Apparantly it wasn't.