Fixed up Player Sprite
Worked a bunch on fixing up the player sprite.  For some reason, it's
doing some really weird things...  I guess I have a bunch of technical
debt that I didn't know about that I had when I started setting up the
first foot so I can't just add the new sprite and call it a day.  But
I'm slowly figuring it out.  It's mostly done, just need to figure out
what the bug is on rotating the foot around the body and do a bunch of
balancing the physics for it.