Fixed bug with Pressure Switch
Finished  the new level and fixed the bug with the pressure switch that
it wasn't reloding.  Mostly the reason it was giving issues is because I
was lazy and just copied the code incorrectly from the action switch
code and didn't comment it up properly.  Oh well.  It works now.
There's a small-ish bug that turns off the switch when you hit it too
hard, but it's only for a fraction of a second.  Not too bad and doesn't
affect things much.  Will probably come back to it eventually.