FreeBox Destruction + Spring Wall
Made the free-move boxes destructable if they hit a hazard.  There's one
thing wrong with it, if I make more than one box, but other than that,
it works well.

Also made a spring-wall prefab that works with the switch and pressure
button.  It makes nice and smooth wall-barring motion.

Found a glitch with the pressure-switch, though.  Gonna have to go back
and fix the pressure switch properly tomorrow.