Finished new level
I finished a version of the last part of the new level.  Unfortunately,
because the LERP block idea didn't work, I had to change up part of it,
and I don't like how it turned out.  I have an idea to make a moving
block act like a LERP block, though.  Im' going to try to implement it
now.